pk: set: Introduce set sub-commands for tree printing
One typical (noob-ish) use for GNU poke is just to parse the content
of a binary file for visual inspection after having defined the binary
structure in a pk file.

Unfortunately the usual output from something like '$<data> @ 0#B'
isn't really user friendly and it is massively difficult to make sense
of the un-spaced wall of text in output.

This patch wants to be an attempt to prettify a bit the displayed data
introducing a new set of commands to tweak how the output data in a
struct is displayed.

Signed-off-by: Carlo Caione <ccaione@baylibre.com>

2020-02-13  Carlo Caione  <ccaione@baylibre.com>

	* testsuite/poke.cmd/set-odepth.pk: new test
	* testsuite/poke.cmd/set-oindent.pk: likewise
	* testsuite/poke.cmd/set-odepth.pk: likewise
	* src/pvm.jitter: new flags in the PVM state struct
	* src/pvm.c: helpers to retrieve tree-related flags
	* src/pvm.h: likewise
	* src/pvm-val.c: support new printing modes
	* src/pk-term.c: new function to print indentation
	* src/pk-term.h: likewise
	* src/pk-set.c: new .set sub-commands
	* doc/poke.texi: updated documentation