maint: Saner diff output
Gnulib's bootstrap script is already set up to prefer the use of
git-merge-changelog as long as we hand-maintain that file (until we
follow through with the thread of switching to gnulib's
gitlog-to-changelog script to autogenerate ChangeLog from our git
history).

Meanwhile, providing some hints to git makes our diff output easier to
parse, with nicer chunk support for *.texi and with nicer overall
patch ordering.

* .gitattributes: New file.
* etc/git.orderfile: New file.
* bootstrap.conf (bootstrap_post_import_hook): Set it up.