ios: Drop Position column from .info ios
Currently, there is only one client of ios_tell() - namely, '.info
ios'.  But in Poke, all actions relating to an ios are given an
explicit offset (even interfaces like dump that don't take an offset
from the user still end up computing a series of offsets before
interacting with the ios).  We don't really have or need a notion of
back-to-back actions where the second can omit the offset and
magically happen at the next bit after the first.  What's more, the
output we were giving was rounded to a byte boundary, and required
extra bookkeeping in the .mem driver to maintain it (the upcoming .nbd
driver would be another driver that has to artificially track current
offset).  While at it, fix some typos.

* src/pk-ios.c (pk_cmd_info_ios, print_info_ios): Simplify.
* src/ios.c (ios_tell): Delete.
* src/ios.h (ios_tell): Likewise.
* doc/poke.texi (info command): Update output.
* testsuite/poke.cmd/file-mode.pk: Likewise.
* testsuite/poke.cmd/file-relative.pk: Likewise.