ios: Change from getchar to pread device interface.
Now that we are passing an offset to every I/O operation, it is not
that much harder to switch from an interface that operates on one byte
at a time to one that operates on a buffer at an offset: pread/pwrite.
For the file driver, there is not a pread counterpart in stdio, so we
still have to fseeko before fread, but as argued in the previous
patch, we're still probably better off sticking to stdio buffers than
additional syscalls were we to rewrite to use pread/pwrite on fds.

The caller, ios.c, still performs most operations one byte at a time,
but that will be cleaned up in subsequent patches.

* src/ios-dev.h (struct ios_dev_if): Replace get_c/put_c with
pread/pwrite.
* src/ios.c (IOS_GET_C_ERR_CHCK, IOS_PUT_C_ERR_CHCK)
(ios_read_string, ios_write_string): Use the new interface.
* src/ios-dev-file.c (ios_dev_file_getc, ios_dev_file_putc):
Rewrite to...
(ios_dev_file_pread, ios_dev_file_pwrite): ...this.
* src/ios-dev-mem.c (ios_dev_mem_getc, ios_dev_mem_putc): Rewrite
to...
(ios_dev_mem_pread, ios_dev_mem_pwrite): ...this.