ios: Prove we don't need seek.
As part of refactoring toward the use of pread rather than fgetc, I
want to validate that the offset passed to every IOS read/write
interaction is the same as the offset previously set with ios->seek.
Since the testsuite still passes, we can then follow this up with a
patch that removes the seek, and rely instead on the offset passed to
the function.

* src/ios-dev.h (struct ios_dev_if): Add offset parameter to get_c
and put_c.
* src/ios.c (IOS_GET_C_ERR_CHCK, IOS_PUT_C_ERR_CHCK)
(ios_read_string, ios_write_string): Pass offset through.
* src/ios-dev-file.c (ios_dev_file_getc, ios_dev_file_putc):
Update signature and add assertion.
* src/ios-dev-mem.c (ios_dev_mem_getc, ios_dev_mem_putc):
Likewise.