The big lib-ization of poke.
This patch splits the implementation in two different components:
- A libpoke.so shared library, containing the PKL, PVM and IOS
  subsystems.  The sources of the library are in lib/.
- A poke program, whose sources are in src/.

There is one pending caveat: when Jitter is used in sub-package mode,
it is available in the form of an archive that contains non-PIC code.
This makes it not possible to link it in libpoke.so in many
architectures, notably x86_64.  Until Jitter fixes this, we link with
the sub-packaged Jitter in the poke program instead.

2020-04-07  Jose E. Marchesi  <jemarch@gnu.org>

	* The big lib-ization.