build: replace libutils with a common sources directory
Placing the little utility code shared between poke and libpoke in
libutils is proving costly (it needs its own gnulib) and complex.  It
is also currently broken.

This patch replaces libutils with a common/ directory that contains
the two files pk-utils.h and pk-utils.c.  Both poke and libpoke add
these files to their SOURCE file lists.

This of course means that both poke and libpoke will be shipping
pk-utils.o, but no big deal: the file is minuscule.

If/when we find that we keep adding stuff to common/ then we can
revisit the idea of having a libutils, maybe even an installable one.

2020-05-09  Jose E. Marchesi  <jemarch@gnu.org>

	* Makefile.am (SUBDIRS): Remove libutils-gl and libutils.
	* configure.ac (AC_CONFIG_FILES): Remove libutils files.
	(AC_CONFIG_MACRO_DIRS): Do not add m4/libutils.
	Do not call libutils_EARLY.
	* libutils/Makefile.am: Remove.
	* bootstrap.conf (bootstrap_post_import_hook): Do not create a
	gl-libutils gnulib.
	* libpoke/Makefile.am (libpoke_la_LIBADD): Remove libutils.la.
	* poke/Makefile.am (poke_LDADD): Likewise.
	* common: Renamed from libutils.
	* HACKING: Update.
	* .gitignore: Update.