pkl: keep data integrity when assigning to struct fields
Setting the value of a struct field may break the data integrity
defined by field constraints and other considerations, such as an
union for which no valid alternatives are found.

This patch makes poke to detect such cases and to not perform the
operation if the data integrity would be endangered.  Example:

  (poke) type Foo = struct { int i : i != 20; }
  (poke) var f = Foo {}
  (poke) f.i = 20
  unhandled constraint violation exception
  (poke) f.i = 21
  (poke) f
  Foo {
    i=21
  }

The same applies to mapped values:

  (poke) var g = Foo @ 0#B
  (poke) g
  Foo {
    i=0xa
  }
  (poke) g.i = 20
  unhandled constraint violation exception
  (poke) g
  Foo {
    i=0xa
  }

Note how the IO space is not modified in case the constraint violation
exception is raised, which is definitely the right thing to do!

2020-11-14  Jose E. Marchesi  <jemarch@gnu.org>

	* libpoke/pkl-insn.def: Define PKL_INSN_SSETI.
	* libpoke/pkl-asm.c (pkl_asm_insn): Handle the SSETI
	macro-instruction.
	(pkl_asm_insn_sseti): New function.
	* libpoke/pkl-asm.pks (sseti): New macro.
	* libpoke/pkl-gen.c (pkl_gen_pr_ass_stmt): Use sseti.
	(pkl_gen_pr_type_struct): Make constructors install constructors
	for their contained anonymous structs.
	* testsuite/poke.pkl/ass-struct-int-1.pk: New test.
	* testsuite/poke.pkl/ass-struct-int-2.pk: Likewise.
	* testsuite/poke.map/ass-map-struct-int-1.pk: Likewise.
	* testsuite/Makefile.am (EXTRA_DIST): Add new tests.