pkl: settle semantics for constructing union values
Unlike structs, union constructors with more than one field
initializer really don't make much sense, like in:

  (poke) type Foo = union { byte b : b > 0; int i; }
  (poke) Foo {b = 2, i = 12}

What kind of Foo are we asking for?  It is not clear.  Until this
patch, the same rules used for constructed structs applied, so the
result would be:

  Foo {
    b=2UB
  }

Now, if we do:

  (poke) var f = Foo {b = 2}
  (poke) f.b = 0

What would we expect?  For the value in `f' to change nature and
become a Foo of kind `i'?  Or for this to be considered as a
constraint error?

In the first case (the nature of the union value changes) we could
expect to get something like:

  (poke) f
  Foo {
  i = 12
  }

But should that really be 12, i.e. the value previously specified in
the initializer, or 0?  Wouldn't we expect this value to impacted by
the coupling in bits of the two fields, if we consider both "start" at
the beginning of the union?

This gets very complicated and obscure, and almost impossible to
implement properly and to understand.  However, this is because we are
thinking about unions like if they were regular structs: they are
not.  This is the wrong way of thinking.

A better approach, which is what this patch implements, is the
following: an union constructor admits either one or zero field
initializers.  Specifying more than one field initializer is a
compile-time error:

  (poke) Foo {b = 2, i = 12}
  <stdin>:1:1: error: union constructors require exactly one field initializer
  Foo {b = 2, i = 12};
  ^~~~~~~~~~~~~~~~~~~

When we specify a field initializer, we are also declaring the kind of
Foo we want to construct, i.e. its "nature":

  (poke) Foo {b = 2}
  (poke) Foo {b = 2}
  Foo {
    b=2UB
  }

Therefore, if we provide an invalid initial value for `b', then we get
a constraint-violation exception:

  (poke) Foo {b = 0}
  unhandled constraint violation exception

The alternative `i' is not considered in this case: we asked for a Foo
of kind `b', and we provided the wrong values for it: we want the
exception.

Once constructed, union values do not change "nature" due to
assignments:

  (poke) var f = Foo {b = 2}
  (poke) f.b = 0
  unhandled constraint violation exception

Note that mapped unions are different in this sense.  When we map an
union on some IO space:

  (poke) var m = Foo @ 0#B

we are not specifying what nature of Foo we want: it all depends on
the contents of the IO space.  Therefore, we could get either a `b' or
an `i', and if the underlying data in the IO space changes, the nature
of the value in `m' may change as well.

Something similar happens when we don't specify a field initializer in
an union constructor:

  (poke) Foo {}
  Foo {
    i=0
  }

Since no initializer was provided, we didn't indicate what kind of Foo
we wanted, and therefore each alternative is tried assuming all fields
have default values, which in the case of integral types is a zero.

Hope all this makes sense XD

2020-11-15  Jose E. Marchesi  <jemarch@gnu.org>

	* libpoke/pkl-typify.c (pkl_typify1_ps_scons): Union constructors
	require exactly one field initializer.
	* libpoke/pkl-gen.pks (struct_constructor): Implement new
	semantics for union construction.
	* testsuite/poke.pkl/scons-56.pk: Adjust test.
	* testsuite/poke.pkl/scons-union-1.pk: Likewise.
	* testsuite/poke.pkl/scons-union-2.pk: Likewise.
	* testsuite/poke.pkl/scons-union-3.pk: LIkewise.
	* testsuite/poke.pkl/scons-union-5.pk: Likewise.
	* testsuite/poke.pkl/scons-union-6.pk: Likewise.
	* testsuite/poke.pkl/scons-union-7.pk: Likewise.
	* testsuite/poke.pkl/scons-union-8.pk: Likewise
	* testsuite/poke.pkl/scons-union-2.pk: Likewise.
	* testsuite/poke.pkl/scons-union-10.pk: New test.
	* testsuite/poke.pkl/scons-union-diag-1.pk: New test.
	* testsuite/poke.pkl/scons-union-diag-2.pk: Likewise.
	* testsuite/poke.pkl/scons-union-9.pk: Likewise.
	* testsuite/poke.pkl/ass-union-int-1.pk: Likewise.
	* testsuite/poke.pkl/ass-union-int-2.pk: Likewise.
	* testsuite/Makefile.am (EXTRA_DIST): Update tests.