pkl: support for array constructors
There were three different ways to create array values in Poke:

1) From an array literal, like `[1,2,3,.[10]=666]'
2) From a mapping, like `int[3] @ 0#B'
3) From the construction of a struct with array fields, like in
   `Packet { } -> Packet { magic = [0UB,0UB], ... }'

However, there was no syntax to construct array values out of the
context of structs, and the array literals are limited, since it is
not possible to express a literal iwth a variable number of elements.

So this patch adds support for array constructors, that use this
syntax:

  ARRAY_TYPE ([INIT_VALUE])

Where ARRAY_TYPE is an array type, or the name of an array type, and
INIT_VALUE is an optional argument that specify the value in the
elements of the array.  If the initial value is not specified, then
default values are used.

Examples:

int[]()               -> []
int[2]()              -> [0,0]
int[2](10)            -> [10,10]
string[3]()           -> ["","",""]
int[][3](int[2](666)) -> [[666,666],[666,666],[666,666]]
Packet[3]()           -> [Packet {...}, Packet {...}, Packet {...}]