pkl: support for mapping and constructing of function types
Up to know trying to map a function type resulted in a compile-time
error:

  (poke) type Func = ()int
  (poke) Func @ 0#B
  <stdin>:1:1: error: specified type cannot be mapped

This made sense, since it is clearly not possible to build function
values from whatever is stored in the IO space.

However, for the sake of orthogonality, and looking to the goal of
supporting storing function values in struct fields, this patch adds
support for mapping functions.

When a function type is mapped, the resulting value is a function that
discards all its arguments and then returns a "default" value of the
proper type, if the function returns a type at all.

Therefore:

(poke) type Func = (int,int) long;
(poke) Func[3] @ 0#B
[#<closure>,#<closure>,#<closure>]
(poke) (Func[3] @ 0#B)[1] (10, 20)
0L

The `size' attribute is also now accepted for function values, and it
always returns 0#B.

2020-11-18  Jose E. Marchesi  <jemarch@gnu.org>

	* libpoke/pkl-ast.c (pkl_ast_type_mappable_p): Function types are
	now mappable.
	* libpoke/pkl-typify.c (pkl_typify1_ps-attr): Function values
	accept the 'size attribute.
	* libpoke/pkl-gen.c (pkl_gen_pr_type_function): New handler.
	(pkl_phase_gen): Install handler.
	* testsuite/poke.pkl/attr-size-14.pk: Likewise.
	* testsuite/poke.pkl/map-diag-1.pk: Remove.
	* testsuite/poke.pkl/unmap-diag-1.pk: Likewise.
	* testsuite/poke.map/func-map-1.pk: New test.
	* testsuite/poke.map/func-map-2.pk: Likewise.
	* testsuite/poke.map/func-map-3.pk: Likewise.
	* testsuite/poke.map/func-map-4.pk: Likewise.
	* testsuite/poke.map/func-map-5.pk: Likewise.
	* testsuite/Makefile.am (EXTRA_DIST): Update.
	* doc/poke.texi (Function Attributes): New section.