build: use gl_PROG_BISON instead of gl_BISON
This is the second attempt of avoiding problems in environments where
YACC is defined.

This patch uses gl_PROG_BISON instead of gl_BISON in configure.ac, and
changes the Makefile.am files in order to not use the automake
implicit rules to build .y files.  Explicit rules are used instead.

2020-11-19  Jose E. Marchesi  <jemarch@gnu.org>

	* configure.ac: Use gl_PROG_BISON instead of gl_BISON.
	* libpoke/Makefile.am (libpoke_la_SOURCES): Replace pkl-tab.y with
	pkl-tab.c.
	(pkl-tab.h pkl-tab.c): New rule to invoke bison.
	(BUILT_SOURCES): Declare pkl-tab.h and pkl-tab.c as built sources.
	(EXTRA_DIST): Distribute pkl-tab.y.
	(MOSTLYCLEANFILES): Add bison generated files.

	* poke/Makefile.am (poke_SOURCES): Replace pk-map-tab.y with
	pk-map-tab.c.
	(BUILT_SOURCES): Add pk-map-tab.c.
	(EXTRA_DIST): Add pk-map-tab.y
	(pk-map-tab.h pk-map-tab.c): New rule to invoke bison.
	(MOSTLYCLEANFILES): Add bison generated files.