build: support for --jitter-srcdir=DIR in `bootstrap'
The patch we sent to gnulib for the `boostrap' script to accept
user-defined hooks to handle command line options has been integrated
upstream.

This patch uses this support it in order to add a new option to our
bootstrap: --jitter-srcdir, which can be used along with --no-git in
order to avoid accessing the internet while bootstrapping.

2020-11-23  Jose E. Marchesi  <jemarch@gnu.org>

	* bootstrap: Updated from latest gnulib.
	* DEV-NEWS: Update.
	* bootstrap.conf (bootstrap_option_hook): New function.
	(bootstrap_print_option_usage_hook): Likewise.
	(bootstrap_post_import_hook): Honor --jitter-srcdir.