libpoke,pickles: Ban anonymous members in unions.
At the Poke level, unions consist of only one member with named
alternates; there is no other way to refer to union members except by
name. This leads to several issues with allowing anonymous union members.

The PVM code generated by struct_printer visits each member in a union
by name, which leads to a segmentation fault when attempting to compile
the printer for unions with anonymous members.

A minimal reproducer:
  (poke) type Foo = union { int; };
  (poke) printf "%v", Foo {};

A second (related) issue is that using unions with anonymous members is
highly problematic. If a union value is created by selecting an
anonymous member, it is a black box. If a union contains multiple
anonymous members, there is no way to distinguish between them.

After discussion, it seems the most sensible solution to these problems
is to forbid anonymous members in unions entirely.

2021-02-08  David Faust  <david.faust@oracle.com>

	* libbpoke/pkl-anal.c: Forbid anonymous members in unions.
	* pickles/btf.pk (type BTF_Type): Give name to anonymous union member.
	* pickles/ctf.pk (type CTF_Type): Likewise.
	* pickles/dwarf-expr.pk (type Dwarf_Op): Likewise.
	* pickles/dwarf-frame.pk (type Dwarf_CFI): Likewise.
	* testsuite/poke.pkl/union-diag-8.pk: New test.
	* testsuite/Makefile.am (EXTRA_DIST): Update.