pickles: Improvements to btf.pk
Several small fixes and improvements to BTF_Type:

Make the offset.bitfield struct an integral struct. This corrects the
ordering of the bitfield_size and bit_offset members, which otherwise
are extracted backwards.

Change several fields representing sizes or offsets to use be offset
types.

Correct the variable-length data portion to also be a BTF_Member[]
for BTF_KIND_STRUCT types.

Also add a new type: struct BTF_Section, which encapsulates all the
information present in a BTF section - header, types and strings.

2021-02-12  David Faust  <david.faust@oracle.com>

	* pickles/btf.pk (BTF_Type) Make several fields `offset`s where
	appropriate. Correct bitfield to use integral struct. Correct variable
	data for BTF_KIND_STRUCT.
	(BTF_Section) New type.