build: do not require a modern AWK at configure time
RAS requires a powerful AWK with support for gensub().  Therefore we
had a check in place in configure.ac that errored out in case no such
an AWK implementation was found in the build host.

However, we are now distributing the RAS-generated .pkc files in the
release tarballs, and therefore this is no longer a concern for users.

This patch moves the check for a RAS suitable AWK implementation from
configure time to build time.  It is the .pkc.pks rule in
libpoke/Makefile.am that emits an error in case the AWK found in the
system is not powerful enough.  Therefore, building from a release
tarball (without modifying any of the .pks files) becomes possible
without having that AWK installed.

2021-02-20  Jose E. Marchesi  <jemarch@gnu.org>

	* configure.ac: Set POKE_AWK and do not error out at configure
	time if no suitable awk is found.
	* libpoke/Makefile.am (.pks.pkc): Check that AWK is powerful
	enough for RAS.  Error out otherwise.