Add installation information for the tarball users.
We want to have an INSTALL file that contains installation instructions for
the tarball users. These instructions are necessarily package specific,
because they need to list the prerequisites or reference the DEPENDENCIES
file.

Unfortunately, both Automake (with strictness option = 'gnu') and 'bootstrap'
both contain code to copy the generic (not package specific) installation
instructions to the file 'INSTALL'. This is not what we want, so we disable
it. (Other packages handle this issue by putting part of the installation
instructions into the README file. This is suboptimal, because this is not
what the README file is for.)

* bootstrap.conf: Arrange to create INSTALL.generic, not INSTALL.
* configure.ac: Pass the Automake option 'foreign'.
* DEPENDENCIES: New file, based on etc/hacking.org.
* INSTALL: New file, based on etc/hacking.org.
* Makefile.am (EXTRA_DIST): Add INSTALL.generic, DEPENDENCIES.
* etc/hacking.org: Remove duplicate documentation.