bootstrap: add the free-posix module to gl-libpoke
This is a workaround for a problem that arises from the fact we are
using a single configure script to configure several different
gnulibs: gl and gl-libpoke.  The gnulib symbol replacement machinery
uses AC_SUBST, which gets applied to the config.h file of both gl and
gl-libpoke.  Therefore, we end with spurious replacements and
link-time errors.

2021-02-22  Jose E. Marchesi  <jemarch@gnu.org>

	* bootstrap.conf (libpoke_modules): Add the free-posix module.