pickles: new pickle asn1-ber.pk
This commit adds a pickle to manipulate ASN-1 data encoded using the
Basic Encoding Rules as specified by the standard ISO/IEC
8825-1:2003.

At the moment it includes support for the generic encoding of data
values, in the type BER_Data_Value.  The interpretation of the value
contents as the different universal types (boolena, integer, etc) is
still to be done.

2021-12-23  Jose E. Marchesi  <jemarch@gnu.org>

	* pickles/asn1-ber.pk: New file.
	* pickles/Makefile.am (dist_pickles_DATA): Add asn1-ber.pk.