pkl, testsuite: support for the EXCOND ?! operator
This operator has two forms:

  EXPRESSION ?! EXCEPTION

and

  COMP_STMT ?! EXCEPTION

i.e. as the first operator it can get an arbitrary expression or a
compound statement.

In both cases, the ?! operator always evaluates to a boolean value.  If
the execution of the expression or statement raises EXCEPTION, the
result of the operation is 0 (false).  Otherwise it is 1 (true).  The
value to which EXPRESSION evaluates is discarded.

Simple example:

 if (1/zero ?! E_div_by_zero)
   print "division by zero\n";

A motivation for this feature is, given an union, an easy way to
determine whether a given alternative is the chosen one without having
to resort to a clumsy try-catch:

  union
  {
    int alt1 : ... ;
    int alt2;

    method _print =
    {
      if (alt1 ?! E_elem)
        pretty_print_alt1;
      else
        pretty_print_alt2;
    }
  }

An example with a statement:

  if ({ do_whatever_io_operation; } ?! E_eof)
    printf "EOF happened\n";

Using the operator in a statement-expression is also useful.  Consider
this idiom we use in pkl-rt-1.pk and other places:

  /* Add the directories in POKE_LOAD_PATH, if defined.  */
  try load_path = getenv ("POKE_LOAD_PATH") + ":" + load_path;
  catch if E_inval { }

This can now be rewritten as:

  /* Add the directories in POKE_LOAD_PATH, if defined.  */
  { load_path = getenv ("POKE_LOAD_PATH") + ":" + load_path } ?! E_inval

2021-12-26  Jose E. Marchesi  <jemarch@gnu.org>

	* libpoke/pkl-ops.def: Entry for PKL_AST_OP_EXCOND.
	* libpoke/pkl-lex.l: Recognize `?!' as EXCOND.
	* libpoke/pkl-tab.y (EXCOND): New token.
	(expression): Add rule for EXCOND.
	* libpoke/pkl-typify.c (pkl_phase_typify1): Install handler.
	(pkl_typify1_ps_op_excond): New handler.
	* libpoke/pkl-gen.c (pkl_gen_pr_op_excond): New handler.
	(pkl_phase_gen): Install handler.
	* testsuite/poke.pkl/excond-diag-1.pk: New test.
	* testsuite/poke.pkl/excond-1.pk: Likewise.
	* testsuite/poke.pkl/excond-2.pk: Likewise.
	* testsuite/poke.pkl/excond-3.pk: Likewise.
	* testsuite/poke.pkl/excond-4.pk: Likewise.
	* testsuite/poke.pkl/excond-5.pk: Likewise.
	* testsuite/Makefile.am (EXTRA_DIST): Add new tests.
	* etc/poke.rec (Exception predicate operator !?): Remove as done.
	* doc/poke.texi (exception-predicate): New section.