poke: new module pk-table.pk
This commit introduces a new Poke facility in the `poke' application,
pk-table.pk, that provides a Pk_Table to ease the output of tabulated
data.  It is basically the Poke counterpart of pk-table.[ch].

The pk_settings_dump (which implements the .set command without
arguments) is also modified to use Pk_Table in order to emit a
well-tabulated relation of all the available settings and their
values.

Note how "execute" terminal hyperlinks are generated in the setting
names to execute ".help SETTING".

2022-01-17  Jose E. Marchesi  <jemarch@gnu.org>

	* poke/pk-table.pk: New file.
	* poke/Makefile.am (dist_pkgdata_DATA): Add pk-table.pk.
	* poke/pk-settings.pk (pk_settings_dump): Use pk_table.