doc: generate settings documentation from poke
The poke settings registry contains the on-line documentation for all
the global settings supported by the tool.

In order to not replicate the same contents in the user manual, this
patch adds a little poke program that dumps the documentation in
Texinfo format.  The manual then @includes this content at the right
section.

2022-01-17  Jose E. Marchesi  <jemarch@gnu.org>

	* doc/settings-to-texi.pk: New file.
	* doc/Makefile.am (poke_TEXINFOS): Add poke-settings.texi.
	(poke-settings.texi): Rule to enerate poke-settings.texi.
	(MAINTAINERCLEANFILES): Add poke-settings.texi.
	(EXTRA_DIST): Likewise.
	* doc/poke.texi (set command): Include poke-settings.texi.