poke: new internal terminal pager
This patch introduces basic paging support in the poke
terminal (pk-term).  A new poke setting `pager' has been introduced to
control whether paging takes effect.  It's default value is `no'.  You
can change this setting (for example in your ~/.pokerc) like this:

  .set pager yes

Note that I am applying this in `master' but not in maint/poke-2.  The
pager works, but I think it needs to be used a little and improved
before being released.  An example of a thing to improve is to
remove (erase) the "--More--" lines when continuing.

2022-01-29  Jose E. Marchesi  <jemarch@gnu.org>

	* poke/pk-cmd-set.c (pk_cmd_set_pager): New function.
	(set_pager_cmd): New variable.
	(pk_cmd_set_init): Add set_pager command.
	* poke/pk-repl.c (pk_repl): Use pager (conditionally) when
	executing commands in the repl.
	* poke/pk-settings.pk: Setting for `pager'.
	* poke/pk-term.c (screen_lines): New variable.
	(screen_cols): Likewise.
	(pager_active_p): Likewise.
	(nlines): Likewise.
	(pk_term_init): Initialize terminal dimensions.
	(pk_term_start_pager): New function.
	(pk_term_stop_pager): Likewise.
	(pk_puts_paged): Likewise.
	(pk_puts): Call pk_puts_paged when necessary.
	(pk_printf): Output via pk_puts.
	* poke/poke.c: New global poke_pager_p.