pickles: asn1-ber: fix issues in long-form length/tag handling
This commit fixes long-form definite length decoding in BER_Length and
BER_Data_Value, so that the octet components are ordered in big-endian
as mentioned in [1], each length octet is not necessarily < 0x80, and
the use of such lengths is not restricted to primitive types:
constructed types such as SEQUENCE can also have definite length.

Besides those this also fixes long-form tag number handling in
BER_Identifier, where it is indicated with 0b11111 not 0b01111.

To reproduce those issues:

  $ wget https://gitlab.com/gnutls/gnutls/-/raw/master/doc/credentials/x509/ca.pem
  $ certtool -i --infile ca.pem --outder --outfile ca.der
  $ ls -al ca.der
  -rw-r--r--. 1 ueno ueno 852 Mar  7 10:18 ca.der
  $ poke ca.der
  (poke) load "asn1-ber.pk"
  (poke) BER_Data_Value @ 0#B
  unhandled constraint violation exception
  (poke) var l = BER_Length @ 1#B
  (poke) l.get
  10246UL#B <-- this should be 848 (= 852 - 4 bytes header)

1. https://en.wikipedia.org/wiki/X.690#Definite_form