pkl,pickles: raise E_constraint when array boundaries types are not satified
Up to now the array mappers and constructors were raising the
E_map_bounds exception when the boundaries of an array type weren't
satisfied (like for example byte[3#b]).

The name of that exception was confusing (constructing doens't involve
mapping) and the extra exception was causing problems with unions.

While fixing that later problem I realized that it is not worth it to
differenciate this case from the more general problem of a constraint
violation: the constraint implicitly defined by the boundary in the
array type.

Therefore this patch.

2022-04-11  Jose E. Marchesi  <jemarch@gnu.org>

	* libpoke/pkl-gen.pks (array_mapper): Raise PVM_E_CONSTRAINT
	instead of PVM_E_MAP_BOUNDS in case of array bound failure.
	(array_constructor): Likewise.
	* libpoke/libpoke.h (PK_EC_MAP_BOUNDS): Remove.
	* libpoke/pkl-rt.pk (E_map_bounds): Likewise.
	* libpoke/pvm.h (PVM_E_MAP_BOUNDS): Likewise.
	(PVM_E_MAP_BOUNDS_NAME): Likewise.
	(PVM_E_MAP_BOUNDS_ESTATUS): Likewise.
	* pickles/diff.pk: Adapt accordingly.
	* testsuite/poke.pkl/scons-union-12.pk: New test.
	* testsuite/poke.pkl/scons-union-13.pk: Likewise.
	* testsuite/poke.pkl/scons-union-14.pk: Likewise.
	* testsuite/poke.map/maps-unions-14.pk: Likewise.
	* testsuite/poke.map/maps-unions-15.pk: Likewise.
	* testsuite/poke.map/maps-unions-16.pk: Likewise.
	* testsuite/Makefile.am (EXTRA_DIST): Add new tests.